refactor(PartData)!: roll PartData methods into ProcessData (makes more sense schematically)
