+++ LotCoMClient/Models/Datasources/ProcessData.cs
@@ -6,7 +6,7 @@ namespace LotCoMClient.Models.Datasources;
 /// <summary>
 /// Provides controlled access to Process data sources.
 /// </summary>
-public class ProcessData 
+public class ProcessData() 
 {
     /// <summary>
     /// Allows interaction with the Process Masterlist Data source file.
@@ -80,16 +80,145 @@ public class ProcessData
     }
 
     /// <summary>
+    /// Retrieves the Process Part list for the specified Process.
+    /// </summary>
+    /// <param name="ProcessFullName"></param>
+    /// <returns></returns>
+    /// <exception cref="ArgumentException"></exception>
+    public List<Part> GetProcessParts(string ProcessFullName) 
+    {
+        // load the Process' data
+        Process Process = Masterlist.GetIndividualProcess(ProcessFullName);
+        // no Part data was read
+        if (Process.Parts.Count < 1) 
+        {
+            throw new ArgumentException($"No Part data found for the Process '{ProcessFullName}'.");
+        } 
+        // return the Part list
+        return Process.Parts;
+    }
+
+    /// <summary>
+    /// Asynchronously retrieves the Process Part list for the specified Process.
+    /// </summary>
+    /// <param name="ProcessFullName">Process FULL Name ("Code-Title") to retrieve Part Data for.</param>
+    /// <returns>A list of Part objects assigned to the Process.</returns>
+    /// <exception cref="ArgumentException"></exception>
+    public async Task<List<Part>> GetProcessPartsAsync(string ProcessFullName) 
+    {
+        // load the Process' data
+        Process Process = await Masterlist.GetIndividualProcessAsync(ProcessFullName);
+        // no Part data was read
+        if (Process.Parts.Count < 1) 
+        {
+            throw new ArgumentException($"No Part data found for the Process '{ProcessFullName}'.");
+        } 
+        // return the Part list
+        return Process.Parts;
+    }
+
+    /// <summary>
+    /// Retrieves and formats ProcessFullName's Part list as a list of Displayable strings.
+    /// </summary>
+    /// <param name="ProcessFullName">Process FULL Name ("Code-Title") to retrieve Part Data for.</param>
+    /// <returns>A List of strings.</returns>
+    public async Task<List<string>> GetDisplayableProcessPartsAsync(string ProcessFullName) 
+    {
+        // retrieve the Process' parts
+        List<Part> ProcessParts = await GetProcessPartsAsync(ProcessFullName);
+        // convert each Part Token into a Displayable string
+        List<string> PartStrings = ProcessParts
+            .Select(x => $"{x.PartNumber}\n{x.PartName}")
+            .ToList();
+        // return the converted list
+        return PartStrings;
+    }
+
+    /// <summary>
+    /// Queries for a Part matching PartNumber in ProcessFullName's Part data.
+    /// </summary>
+    /// <param name="ProcessFullName">The FULL Name ("Code-Title") of the Process to query from.</param>
+    /// <param name="PartNumber">The Part Number to query for within ProcessFullName's data.</param>
+    /// <returns>A JToken object containing the Part data for PartNumber.</returns>
+    /// <exception cref="ArgumentException"></exception>
+    /// <exception cref="FormatException"></exception>
+    public Part GetProcessPartData(string ProcessFullName, string PartNumber) 
+    {
+        // retrieve the Process' Part list
+        List<Part> ProcessParts = GetProcessParts(ProcessFullName);
+        // no Part data for this Process
+        if (ProcessParts.Count == 0) 
+        {
+            throw new ArgumentException($"No Part data has been assigned to Process '{ProcessFullName}'.");
+        }
+        // attempt to access the specific Part
+        Part? SelectedPart;
+        try 
+        {
+            SelectedPart = ProcessParts
+                .Where(x => x.PartNumber
+                .Equals(PartNumber))
+                .First();
+        // Part was not found in the Process' Part list
+        } 
+        catch 
+        {
+            throw new ArgumentException($"Part '{PartNumber}' not found assigned to Process '{ProcessFullName}'.");
+        }
+        return SelectedPart;
+    }
+
+    /// <summary>
+    /// Asynchronously queries for a Part matching PartNumber in ProcessFullName's Part data.
+    /// </summary>
+    /// <param name="ProcessFullName">The FULL Name ("Code-Title") of the Process to query from.</param>
+    /// <param name="PartNumber">The Part Number to query for within ProcessFullName's data.</param>
+    /// <returns>A JToken object containing the Part data for PartNumber.</returns>
+    /// <exception cref="ArgumentException"></exception>
+    /// <exception cref="FormatException"></exception>
+    public async Task<Part> GetProcessPartDataAsync(string ProcessFullName, string PartNumber) 
+    {
+        // perform the query on a new CPU thread
+        Part PartData = await Task.Run(async () => 
+        {
+            // retrieve the Process' Part list
+            List<Part> ProcessParts = await GetProcessPartsAsync(ProcessFullName);
+            // no Part data for this Process
+            if (ProcessParts.Count == 0) 
+            {
+                throw new ArgumentException($"No Part data has been assigned to Process '{ProcessFullName}'.");
+            }
+            // attempt to access the specific Part
+            Part? SelectedPart;
+            try 
+            {
+                SelectedPart = ProcessParts
+                    .Where(x => x.PartNumber
+                    .Equals(PartNumber))
+                    .First();
+            // Part was not found in the Process' Part list
+            } 
+            catch 
+            {
+                throw new ArgumentException($"Part '{PartNumber}' not found assigned to Process '{ProcessFullName}'.");
+            }
+            return SelectedPart;
+        });
+        // return the queried Part data
+        return PartData;
+    }
+
+    /// <summary>
     /// Provides ProcessData controlled access to the Process Masterlist data source.
     /// </summary>
-    private class ProcessMasterlist 
+    private class ProcessMasterlist()
     {
         private const string Path = "\\\\144.133.122.1\\Lot Control Management\\Database\\process_control\\_process_masterlist.json";
         
         /// <summary>
         /// Contains the full JObject object produced by the last LoadDataAsync call. 
         /// </summary>
-        private JObject? LastRead;
+        private JObject? LastRead = null;
 
         /// <summary>
         /// Asynchronously loads the data from the Process Masterlist data source. Stores this data in the LastRead property.
@@ -98,7 +227,6 @@ public class ProcessData
         /// <exception cref="JsonException"></exception>
         private async Task<JObject> LoadDataAsync() 
         {
-            Console.WriteLine("Reading...");
             // read the masterlist file
             LastRead = JObject.Parse(await File.ReadAllTextAsync(Path));
             return LastRead;
@@ -111,7 +239,6 @@ public class ProcessData
         /// <exception cref="JsonException"></exception>
         private JObject LoadData() 
         {
-            Console.WriteLine("Reading...");
             // read the masterlist file
             LastRead = JObject.Parse(File.ReadAllText(Path));
             return LastRead;
@@ -382,6 +509,7 @@ public class ProcessData
         {
             if (LastRead is null) 
             {
+                Console.WriteLine($"LastRead: {LastRead is null}. Reading from GetIndividualProcessAsync()...");
                 // load the data from the Masterlist
                 await LoadDataAsync();
             }
